Collapsed mask into gray image, readme, comments
Added the training files in the resources folder.
Comments on many of the public methods of important classes
Unit tests for the algorithm and computer vision portions of the system
Temporarily disabled a part of the color isolation algorithm as it was
giving me trouble. Something changed in refactoring and it needs to be
solved.