Maven changes, opencv loading fail-fast + windows
Changed the apache CLI library to use a library on maven central that
anyone will be able to get access to. It is the same as the previous
one, just on the net.

All OpenCV loading now takes place from OpenCV#load(), and is fail fast.
Includes support for Windows with message indicating what must be done
for Windows support to be set up correctly (moving the \lib directory
and its contents).